separate main class in different files
