+++ src/PerformanceMonitor/PerformanceMonitor.BlockProperties.cs
@@ -1,0 +1,94 @@
+using Neo.Ledger;
+using Neo.Network.P2P.Payloads;
+using System;
+
+namespace Neo.Plugins
+{
+    partial class PerformanceMonitor
+    {
+        /// <summary>
+        /// Process "block time" command
+        /// Prints the block time in seconds of the given block index or block hash
+        /// </summary>
+        private bool OnBlockTimeCommand(string[] args)
+        {
+            if (args.Length != 3)
+            {
+                return false;
+            }
+            else
+            {
+                string blockId = args[2];
+                Block block = null;
+
+                if (UInt256.TryParse(blockId, out var blockHash))
+                {
+                    block = Blockchain.Singleton.GetBlock(blockHash);
+                }
+                else if (uint.TryParse(blockId, out var blockIndex))
+                {
+                    block = Blockchain.Singleton.GetBlock(blockIndex);
+                }
+
+                if (block == null)
+                {
+                    Console.WriteLine("Block not found");
+                }
+                else
+                {
+                    ulong time = block.GetTime();
+
+                    Console.WriteLine($"Block Hash: {block.Hash}");
+                    Console.WriteLine($"      Index: {block.Index}");
+                    Console.WriteLine($"      Time: {time / 1000} seconds");
+                }
+
+                return true;
+            }
+        }
+
+        /// <summary>
+        /// Process "block avgtime" command
+        /// Prints the average time in seconds the latest blocks are active
+        /// </summary>
+        private bool OnBlockAverageTimeCommand(string[] args)
+        {
+            if (args.Length > 3)
+            {
+                return false;
+            }
+            else
+            {
+                uint desiredCount = 1000;
+                if (args.Length == 3)
+                {
+                    if (!uint.TryParse(args[2], out desiredCount))
+                    {
+                        Console.WriteLine("Invalid parameter");
+                        return true;
+                    }
+
+                    if (desiredCount < 1)
+                    {
+                        Console.WriteLine("Minimum 1 block");
+                        return true;
+                    }
+
+                    if (desiredCount > 10000)
+                    {
+                        Console.WriteLine("Maximum 10000 blocks");
+                        return true;
+                    }
+                }
+
+                using (var snapshot = Blockchain.Singleton.GetSnapshot())
+                {
+                    var averageInSeconds = snapshot.GetAverageTimePerBlock(desiredCount) / 1000;
+                    Console.WriteLine(averageInSeconds.ToString("Average time/block: 0.00 seconds"));
+                }
+
+                return true;
+            }
+        }
+    }
+}
